Replaced the Main Logo image. Added a Banner image version of the Logo. Referenced this on all pages
Added the data to be included in the table on the Hills page. Columns Hill Name, Height (m), County and Completed. Completed column includes a checkbox. This needs to be updated to allow the checkboxe states to be saved when updated.

Issues with labels on this table, eventually resolved by calling each label the same value as the hill name