Created variables to hold the data in the geolocation object: latitude longitude accuracy timestamp datetime
Set the new variables values to those in the geolocation object.
Displayed the output using the new variables.
Converted the UNIX timestamp for geolocation to a user readable datetime - UTC string was chosedn because it shows all of the relevant data without including inecessary information (such as time zone in detail)