Split the showPosition method into getData and showPosition. This allows for separation between the functionality of each method. getData method sets all of the newly created variables to hold the values of the geolocation data showPosition method now only displays the information to the user.
Functionality testing on the co-ordinates has been completed. Compared the values returned in this JS code to an independant app on a different device. The app used was called 'Coordinates - GPS Formatter' and this was run on an iPhone 7 running iOS 13.3. The phone was next to the computer at the time of testing.

Testing shows the co-ordinates are correct within the accuracy threshold provided.